debaide-main/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.0
python-multipart==0.0.12
google-generativeai==0.8.3
openai-whisper
boto3==1.35.70
python-dotenv==1.0.1
